+++ redditrepostsleuth/submonitorsvc/sub_monitor_svc_celery.py
@@ -25,14 +25,15 @@ if __name__ == '__main__':
     event_logger = EventLogging(config=config)
     uowm = SqlAlchemyUnitOfWorkManager(get_db_engine(config))
     response_builder = ResponseBuilder(uowm)
-    dup = DuplicateImageService(uowm, event_logger, config=config)
-    reddit = RedditManager(get_reddit_instance(config))
+    reddit = get_reddit_instance(config)
+    reddit_manager = RedditManager(reddit)
+    dup = DuplicateImageService(uowm, event_logger, reddit, config=config)
     monitor = SubMonitor(
         dup,
         uowm,
-        reddit,
+        reddit_manager,
         response_builder,
-        ResponseHandler(reddit, uowm, event_logger, source='submonitor', live_response=config.live_responses),
+        ResponseHandler(reddit_manager, uowm, event_logger, source='submonitor', live_response=config.live_responses),
         event_logger=event_logger,
         config=config
     )