+++ changelog.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [UNRELEASED]
 * [bugfix] - Matches from current month were getting dropped if IDs overlapped historical images
 * [backend] - Added API to support repostsleuth.com
+* [feature] - Bot now checks if a post is removed from Reddit, in addition to checking if the image has been removed
 
 
 ## [0.1.8] - 8/29/2020